fix(013): high-res float encoder for velocity loop (throb: L CV 0.2->0.04)
Velocity loop now differentiates a float-resolution encoder position instead of
one truncated to whole mm (Motor::readEncoderMmF; _prevEnc float). The 1mm
position quantization was ~17mm/s velocity noise at the ~58ms loop that the
left-wheel PI chased into a throb. Combined with the per-tick readSpeed removal,
left-wheel velocity CV dropped from ~0.9 -> ~0.04 (objective; analyzer averages
~200ms so a faint sub-200ms residual remains, per stakeholder observation).

+++ source/control/MotorController.h
@@ -101,8 +101,8 @@ private:
     float _tgtRMms;
 
     // Cached encoder readings from the most recent tick() call.
-    int32_t _prevEncL;   // mm at start of last tick
-    int32_t _prevEncR;
+    float   _prevEncL;   // mm at start of last tick (float — high-res for velocity)
+    float   _prevEncR;
     float   _actualVelL; // mm/s computed in tick()
     float   _actualVelR;
     float   _encLMm;     // current encoder position (mm), updated in tick()
